Hopefully finished model building just need to test.

+++ DMS/app/src/main/java/com/example/kbb12/dms/bolusInsulinModelBuilder/controller/BreakCarbsListener.java
@@ -28,7 +28,7 @@ public class BreakCarbsListener implements TextWatcher {
     @Override
     public void afterTextChanged(Editable s) {
         if(s.toString().length()>0) {
-            model.setBreakCarbs(Integer.parseInt(s.toString()));
+            model.setBreakCarbs(Double.parseDouble(s.toString()));
         }else{
             model.setBreakCarbs(null);
         }
