+++ DMS/app/src/main/java/com/example/kbb12/dms/bolusInsulinModelBuilder/controller/DinCarbsListener.java
@@ -28,7 +28,7 @@ public class DinCarbsListener implements TextWatcher {
     @Override
     public void afterTextChanged(Editable s) {
         if(s.toString().length()>0){
-            model.setDinCarbs(Integer.parseInt(s.toString()));
+            model.setDinCarbs(Double.parseDouble(s.toString()));
         }else{
             model.setDinCarbs(null);
         }
