+++ DMS/app/src/main/java/com/example/kbb12/dms/bolusInsulinModelBuilder/controller/DinInsulinListener.java
@@ -28,7 +28,7 @@ public class DinInsulinListener implements TextWatcher {
     @Override
     public void afterTextChanged(Editable s) {
         if(s.toString().length()>0) {
-            model.setDinInsulin(Integer.parseInt(s.toString()));
+            model.setDinInsulin(Double.parseDouble(s.toString()));
         }else{
             model.setDinInsulin(null);
         }
